+++ fmpy/__init__.py
@@ -5,7 +5,7 @@ import os
 from ctypes import *
 import _ctypes
 
-__version__ = '0.2.10'
+__version__ = '0.2.11'
 
 
 # determine the platform
@@ -179,7 +179,7 @@ def extract(filename, unzipdir=None):
         unzipdir = mkdtemp()
 
     # expand the 8.3 paths on windows
-    if sys.platform.startswith('win'):
+    if sys.platform.startswith('win') and '~' in unzipdir:
         import win32file
         unzipdir = win32file.GetLongPathName(unzipdir)